// src/types/user.ts
export interface User {
    id: number;
    tckno: string;
    name: string;
    surname: string;
    email: string;
    phone1: string;
    phone2: string;
    userName: string;
    department: string;
    departmentId: number;
    title: string;
    rolName: string;
    rolId: number;
    cinsiyet: string;
    image: string;
  }